+++ cypress/e2e/tests/adminPanel/c399_analiticMenuFunctionality.cy.ts
